rust: move Key, PressedKey traits to key mod
