+++ src/key/mod.rs
@@ -1,9 +1,37 @@
+use core::fmt::Debug;
+
 use crate::input;
 
 pub mod simple;
 pub mod tap_hold;
 
 pub mod composite;
+
+pub trait Key {
+    type PressedKey: PressedKey<Key = Self, Event = Self::Event> + Debug;
+    type Event: Copy + Debug + Ord;
+
+    fn new_pressed_key(
+        keymap_index: u16,
+        key_definition: &Self,
+    ) -> (Self::PressedKey, Option<ScheduledEvent<Self::Event>>);
+}
+
+pub trait PressedKey {
+    type Event;
+    type Key: Key;
+    fn handle_event(
+        &mut self,
+        key_definition: &Self::Key,
+        event: Event<Self::Event>,
+    ) -> impl IntoIterator<Item = Event<Self::Event>>;
+    fn key_code(&self, key_definition: &Self::Key) -> Option<u8>;
+}
+
+#[allow(unused)]
+pub enum EventError {
+    UnmappableEvent,
+}
 
 #[derive(Debug, Clone, Copy, PartialEq, PartialOrd, Eq, Ord)]
 pub enum Event<T> {
